refactor(docx): isolate textbox anchor compatibility
Move the visible text-body anchor extent policy behind the reviewed compatibility boundary and register its focused regression evidence. The retained story and rendered geometry are unchanged.\n\nVerified with textbox layout tests, compatibility evidence checks, the final layout boundary, DOCX typecheck, and diff checks.\n\nCo-Authored-By: Codex GPT-5.6 Sol <noreply@openai.com>
